# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API that allows clients to communicate with drones (i.e. **dispatch controller**)",
        "title": "github template restapi",
        "contact": {
            "name": "Daniel Mena and Kmilo Denis Glez",
            "url": "https://github.com/dani-fmena and https://github.com/kmilodenisglez",
            "email": "kmilo.denis.glez@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/database/populate": {
            "post": {
                "description": "Populate the database with the following data:\r\n\r\n`two` users for authentication:\r\n\r\n```json\r\n[\r\n  {\r\n    \"passphrase\": \"0b14d501a594442a01c6859541bcb3e8164d183d32937b851835442f69d5c94e\",\r\n    \"username\": \"richard.sargon@meinermail.com\",\r\n    \"name\": \"Richard Sargon\"\r\n  },\r\n  {\r\n    \"passphrase\": \"6cf615d5bcaac778352a8f1f3360d23f02f34ec182e259897fd6ce485d7870d4\",\r\n    \"username\": \"tom.carter@meinermail.com\",\r\n    \"name\": \"Tom Carter\"\r\n  }\r\n]\r\n```\r\n\r\n```text\r\npassword for 'richard.sargon@meinermail.com' user is: password1\r\npassword for 'tom.carter@meinermail.com' user is: password2\r\n```\r\n\r\n`ten` drones:\r\n\r\n\r\nModel enum for a Drone:\r\n```text\r\n0 =\u003e Lightweight\r\n1 =\u003e Middleweight\r\n2 =\u003e Cruiserweight\r\n3 =\u003e Heavyweight\r\n\r\n```\r\n\r\nState enum for a Drone:\r\n```text\r\n0 =\u003e IDLE\r\n1 =\u003e LOADING\r\n2 =\u003e LOADED\r\n3 =\u003e DELIVERING\r\n4 =\u003e DELIVERED\r\n5 =\u003e RETURNING\r\n```\r\n\r\ngenerated drone collection:\r\n```json\r\n[\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174001\",\"model\":2,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174002\",\"model\":1,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174003\",\"model\":3,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174004\",\"model\":1,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174005\",\"model\":3,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174006\",\"model\":0,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174007\",\"model\":2,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174008\",\"model\":3,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174009\",\"model\":0,...},\r\n  {\"serialNumber\":\"123e4567-e89b-12d3-a456-426614174010\",\"model\":0,...}\r\n]\r\n```\r\n\r\nseven (7) medications:\r\n```json\r\n[\r\n  {\"name\":\"a random string\",\"weight\":115,\"code\":\"a random code\",\"image\":\"ZmFrZV9pbWFnZQ==\"}, \r\n  {\"name\":\"a random string\",\"weight\":10,\"code\":\"a random code\",\"image\":\"ZmFrZV9pbWFnZQ==\"},\r\n  {\"name\":\"a random string\",\"weight\":210,\"code\":\"a random code\",\"image\":\"ZmFrZV9pbWFnZQ==\"},\r\n  {\"name\":\"a random string\",\"weight\":34,\"code\":\"a random code\",\"image\":\"ZmFrZV9pbWFnZQ==\"}\r\n  ...\r\n]\r\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "database"
                ],
                "summary": "Populate the database with fake data",
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/drones": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all drones or you can filter by status\r\n\r\n\r\nModel enum for a Drone:\r\n```text\r\n0 =\u003e Lightweight\r\n1 =\u003e Middleweight\r\n2 =\u003e Cruiserweight\r\n3 =\u003e Heavyweight\r\n\r\n```\r\n\r\nState enum for a Drone:\r\n```text\r\n0 =\u003e IDLE\r\n1 =\u003e LOADING\r\n2 =\u003e LOADED\r\n3 =\u003e DELIVERING\r\n4 =\u003e DELIVERED\r\n5 =\u003e RETURNING\r\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drones"
                ],
                "summary": "Get drones",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5
                        ],
                        "type": "integer",
                        "description": "drone state",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Drone"
                            }
                        }
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register or update a drone in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drones"
                ],
                "summary": "Registers a new drone, also updates a previously inserted drone",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Drone data",
                        "name": "drone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestDrone"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/drones/{serialNumber}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a drone by the serialNumber",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drones"
                ],
                "summary": "Get a drone by serialNumber",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "Serial number of a drone",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Drone"
                        }
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/medications": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return the medication items\r\n\r\nTake the code medication item as the unique identifier.\r\n\r\nExample Response body:\r\n````json\r\n[\r\n  {\r\n    \"name\": \"Hcj6VWVAzP1g\",\r\n    \"weight\": 700,\r\n    \"code\": \"OJ02IBPL0U\",\r\n    \"image\": \"ZmFrZV9pbWFnZQ==\"\r\n  },\r\n  {\r\n    \"name\": \"CANCAPITAL\",\r\n    \"weight\": 490,\r\n    \"code\": \"60FJFAQ93H\",\r\n    \"image\": \"ZmFrZV9pbWFnZQ==\"\r\n  },...\r\n]\r\n````\r\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medications"
                ],
                "summary": "Get medications",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Medication"
                            }
                        }
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        },
        "/medications/items/{serialNumber}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Checking loaded medication items for a given drone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medications"
                ],
                "summary": "Checking loaded medication items for a given drone",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "Serial number of a drone",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Load or Update a drone with medication items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medications"
                ],
                "summary": "Load a drone with medication items",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "Serial number of a drone",
                        "name": "serialNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medication item codes' collection",
                        "name": "medicationItemCodes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "err.processing_param",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "401": {
                        "description": "err.unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "500": {
                        "description": "err.database_related",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    },
                    "504": {
                        "description": "err.network",
                        "schema": {
                            "$ref": "#/definitions/dto.Problem"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Drone": {
            "type": "object",
            "properties": {
                "batteryCapacity": {
                    "type": "number"
                },
                "model": {
                    "type": "integer"
                },
                "serialNumber": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "weightLimit": {
                    "type": "number"
                }
            }
        },
        "dto.Medication": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "we assume that the code is unique",
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Some error details"
                },
                "status": {
                    "type": "integer",
                    "example": 503
                },
                "title": {
                    "type": "string",
                    "example": "err_code"
                }
            }
        },
        "dto.RequestDrone": {
            "type": "object",
            "properties": {
                "batteryCapacity": {
                    "type": "number"
                },
                "model": {
                    "type": "integer"
                },
                "serialNumber": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
...
    }
}